requirements.txt
astropy==7.2.0
astropy-iers-data==0.2026.3.2.0.47.4
astroquery==0.4.11
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dash==4.1.0
Flask==3.1.3
html5lib==1.1
idna==3.11
importlib_metadata==9.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
keyring==25.7.0
MarkupSafe==3.0.3
more-itertools==10.8.0
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
plotly==6.6.0
psycopg2==2.9.11
pyerfa==2.0.1.5
python-dotenv==1.2.2
pyvo==1.8.1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
retrying==1.4.2
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.7
zipp==3.23.0
